// projects/frontend/src/app/entities/organisation-form/interfaces/organisation.interface.ts
export interface IOrganisation {
  organisationName: string;
  organisationLegalEntity: string;
  mainContactPerson: string;
  contactEmail: string;
  organisationAddress: string;
  role: string;
  roles: string[];
  avatar: string;
  street: string,
  buildingName: string,
  zipCode: string,
  city: string,
  country: string;
};